+++ router/batchrouter/asyncdestinationmanager/snowpipestreaming/channel.go
@@ -12,7 +12,10 @@ import (
 	whutils "github.com/rudderlabs/rudder-server/warehouse/utils"
 )
 
-func (m *Manager) prepareChannelResponse(
+// initializeChannelWithSchema creates a new channel for the given table if it doesn't exist.
+// If the channel already exists, it checks for new columns and adds them to the table.
+// It returns the channel response after creating or recreating the channel.
+func (m *Manager) initializeChannelWithSchema(
 	ctx context.Context,
 	destinationID string,
 	destConf *destConfig,
@@ -36,6 +39,38 @@ func (m *Manager) prepareChannelResponse(
 		}
 	}
 	return channelResponse, nil
+}
+
+func findNewColumns(eventSchema, snowPipeSchema whutils.ModelTableSchema) []whutils.ColumnInfo {
+	var newColumns []whutils.ColumnInfo
+	for column, dataType := range eventSchema {
+		if _, exists := snowPipeSchema[column]; !exists {
+			newColumns = append(newColumns, whutils.ColumnInfo{
+				Name: column,
+				Type: dataType,
+			})
+		}
+	}
+	return newColumns
+}
+
+// addColumns adds columns to a Snowflake table one at a time, as ALTER TABLE does not support IF NOT EXISTS with multiple columns.
+func (m *Manager) addColumns(ctx context.Context, namespace, tableName string, columns []whutils.ColumnInfo) error {
+	m.logger.Infon("Adding columns", logger.NewStringField("table", tableName))
+
+	snowflakeManager, err := m.createSnowflakeManager(ctx, namespace)
+	if err != nil {
+		return fmt.Errorf("creating snowflake manager: %w", err)
+	}
+	defer func() {
+		snowflakeManager.Cleanup(ctx)
+	}()
+	for _, column := range columns {
+		if err = snowflakeManager.AddColumns(ctx, tableName, []whutils.ColumnInfo{column}); err != nil {
+			return fmt.Errorf("adding column: %w", err)
+		}
+	}
+	return nil
 }
 
 // createChannel creates a new channel for importing data to Snowpipe.
@@ -201,7 +236,7 @@ func (m *Manager) createSnowflakeManager(ctx context.Context, namespace string) 
 	}
 	modelWarehouse.Destination.Config["useKeyPairAuth"] = true // Since we are currently only supporting key pair auth
 
-	sf, err := manager.New(whutils.SnowpipeStreaming, m.conf, m.logger, m.statsFactory)
+	sf, err := manager.New(whutils.SnowpipeStreaming, m.appConfig, m.logger, m.statsFactory)
 	if err != nil {
 		return nil, fmt.Errorf("creating snowflake manager: %w", err)
 	}
