--- router/batchrouter/asyncdestinationmanager/snowpipestreaming/transform.go
@@ -1,10 +1,0 @@
-package snowpipestreaming
-
-import (
-	"github.com/rudderlabs/rudder-server/jobsdb"
-	"github.com/rudderlabs/rudder-server/router/batchrouter/asyncdestinationmanager/common"
-)
-
-func (m *Manager) Transform(job *jobsdb.JobT) (string, error) {
-	return common.GetMarshalledData(string(job.EventPayload), job.JobID)
-}
